menus done

+++ src/data_innit.c
@@ -33,7 +33,7 @@ char *mkrndstr(size_t length) { // const size_t length, supra
 	    if (randomString) {
 	        int l = (int) (sizeof(charset) -1); // (static/global, could be const or #define SZ, would be even better)
 	        int key;  // one-time instantiation (static/global would be even better)
-	        for (int n = 0;n < length;n++) {        
+	        for (int n = 0;n < (int)length;n++) {        
 	            key = rand() % l;   // no instantiation, just assignment, no overhead from sizeof
 	            randomString[n] = charset[key];
 	        }
